Moved all dashboard files into Stores folder
